Splitting operators into hpp/cpp and updating its recipe
