+++ cpp/tests/algebra/CMakeLists.txt
@@ -30,6 +30,7 @@ add_executable(operators
 
 target_link_libraries(operators
     PRIVATE GTest::gtest_main
-    PRIVATE algebra)
+    PRIVATE algebra
+    PRIVATE log)
 
 gtest_add_tests(TARGET operators)
